Step 22: Renames Rollback.cc/.h -> ConfigManager.cc/.h
