+++ mgmt/ConfigManager.h
